Badge to show you're free of known vulnerabilities
bandiera-client-node has no vulnerabilities, which is awesome! 

This PR adds a badge that shows you're free of known vulnerabilities. Note that the badge works off the latest master branch, so it'll show red until you actually merge this in.

Adding the badge will show others that you care about security, and that they should care, too. 
(it will also help to spread the word about Snyk, which would help us a lot!).